Add ability to select strategy
