+++ .user.cfg.example
@@ -7,3 +7,4 @@ tld=com
 hourToKeepScoutHistory=1
 scout_multiplier=5
 scout_sleep_time=5
+strategy=DefaultStrategy
