+++ binance_trade_bot/crypto_trading.py
@@ -1,12 +1,12 @@
 #!python3
 import time
 
-from .auto_trader import AutoTrader
 from .binance_api_manager import BinanceAPIManager
 from .config import Config
 from .database import Database
 from .logger import Logger
 from .scheduler import SafeScheduler
+from .strategies import get_strategy
 
 
 def main():
@@ -16,7 +16,11 @@ def main():
     config = Config()
     db = Database(logger, config)
     manager = BinanceAPIManager(config, db, logger)
-    trader = AutoTrader(manager, db, logger, config)
+    strategy = get_strategy(config.STRATEGY)
+    if strategy is None:
+        logger.error("Invalid strategy name")
+        return
+    trader = strategy(manager, db, logger, config)
 
     logger.info("Creating database schema if it doesn't already exist")
     db.create_database()
@@ -24,8 +28,7 @@ def main():
     db.set_coins(config.SUPPORTED_COIN_LIST)
     db.migrate_old_state()
 
-    trader.initialize_trade_thresholds()
-    trader.initialize_current_coin()
+    trader.initialise()
 
     schedule = SafeScheduler(logger)
     schedule.every(config.SCOUT_SLEEP_TIME).seconds.do(trader.scout).tag("scouting")
