+++ binance_trade_bot/strategies/README.md
@@ -1,0 +1,2 @@
+# Strategies
+## DefaultStrategy
