+++ binance_trade_bot/strategies/__init__.py
@@ -1,0 +1,5 @@
+from .default import DefaultStrategy
+
+
+def get_strategy(name):
+    return next((S for S in [DefaultStrategy] if S.__name__ == name), None)
